user-input
I finished making the user-input-form, and the handle submit function. From the handle submit function I made an api call from the front end of the user-input component to google's api to  translate the typed word by the user. Then I started the fetch from the response from googles api to the backend to be able to save the users words into supabase. The last thing I have to do is make an api route to connect the flashcards to the flashcard table in supabase